Refactored containers, distribution methods, and corrected imports in all dependent files
