+++ pacai/agents/capture/capture.py
@@ -6,8 +6,9 @@ import abc
 
 from pacai.agents.base import BaseAgent
 from pacai.ui import captureGraphicsDisplay
+from pacai.util import counter
 from pacai.util import distanceCalculator
-from pacai.util import containers
+from pacai.util import util
 
 class CaptureAgent(BaseAgent):
     """
@@ -115,7 +116,7 @@ class CaptureAgent(BaseAgent):
 
         myState = gameState.getAgentState(self.index)
         myPos = myState.getPosition()
-        if myPos != distanceCalculator.nearestPoint(myPos):
+        if myPos != util.nearestPoint(myPos):
             # We're halfway from one position to the next.
             return gameState.getLegalActions(self.index)[0]
         else:
@@ -241,7 +242,7 @@ class CaptureAgent(BaseAgent):
         Overlays a distribution over positions onto the pacman board that represents
         an agent's beliefs about the positions of each agent.
 
-        The arg distributions is a tuple or list of containers.Counter objects, where the i'th
+        The arg distributions is a tuple or list of counters.Counter objects, where the i'th
         Counter has keys that are board positions (x,y) and values that encode the probability
         that agent i is at (x,y).
 
@@ -254,11 +255,11 @@ class CaptureAgent(BaseAgent):
         dists = []
         for dist in distributions:
             if (dist is not None):
-                if not isinstance(dist, containers.Counter):
+                if not isinstance(dist, counter.Counter):
                     raise Exception("Wrong type of distribution")
                 dists.append(dist)
             else:
-                dists.append(containers.Counter())
+                dists.append(counter.Counter())
 
         if (self.display is not None and 'updateDistributions' in dir(self.display)):
             self.display.updateDistributions(dists)
