+++ pacai/agents/capture/defense.py
@@ -1,6 +1,6 @@
 from pacai.agents.capture.reflex import ReflexCaptureAgent
 from pacai.core.game import Directions
-from pacai.util import containers
+from pacai.util import counter
 
 class DefensiveReflexAgent(ReflexCaptureAgent):
     """
@@ -13,7 +13,7 @@ class DefensiveReflexAgent(ReflexCaptureAgent):
         super().__init__(index)
 
     def getFeatures(self, gameState, action):
-        features = containers.Counter()
+        features = counter.Counter()
         successor = self.getSuccessor(gameState, action)
 
         myState = successor.getAgentState(self.index)
