+++ pacai/agents/capture/offense.py
@@ -1,5 +1,5 @@
 from pacai.agents.capture.reflex import ReflexCaptureAgent
-from pacai.util import containers
+from pacai.util import counter
 
 class OffensiveReflexAgent(ReflexCaptureAgent):
     """
@@ -12,7 +12,7 @@ class OffensiveReflexAgent(ReflexCaptureAgent):
         super().__init__(index)
 
     def getFeatures(self, gameState, action):
-        features = containers.Counter()
+        features = counter.Counter()
         successor = self.getSuccessor(gameState, action)
         features['successorScore'] = self.getScore(successor)
