+++ pacai/agents/learning/reinforcement.py
@@ -3,7 +3,7 @@ import logging
 import time
 
 from pacai.agents.learning.value import ValueEstimationAgent
-from pacai.util import containers
+from pacai.util import counter
 
 class ReinforcementAgent(ValueEstimationAgent):
     """
