+++ pacai/agents/learning/value.py
@@ -1,7 +1,7 @@
 import abc
 
 from pacai.agents.base import BaseAgent
-from pacai.util import containers
+from pacai.util import counter
 
 class ValueEstimationAgent(BaseAgent):
     """
