+++ pacai/util/debug.py
@@ -1,0 +1,7 @@
+def pause():
+    """
+    Pauses the output stream awaiting user feedback.
+    """
+
+    print("<Press enter/return to continue>")
+    input()
