+++ tests/test_util_containers.py
@@ -1,6 +1,8 @@
 import unittest
 
-from pacai.util import containers
+from pacai.util import priority_queue
+from pacai.util import queue
+from pacai.util import stack
 
 """
 This is a test class to assess the functionality of the data structures
@@ -9,47 +11,47 @@ and the Priority Queue.
 """
 class UtilTest(unittest.TestCase):
     def test_queue(self):
-        queue = containers.Queue()
-        self.assertTrue(queue.isEmpty())
+        q = queue.Queue()
+        self.assertTrue(q.isEmpty())
 
         val_list = [x for x in range(1, 10)]
         for val in val_list:
-            queue.push(val)
-        self.assertFalse(queue.isEmpty())
-        self.assertEquals(len(val_list), len(queue))
+            q.push(val)
+        self.assertFalse(q.isEmpty())
+        self.assertEquals(len(val_list), len(q))
 
         # Test Queue for FIFO functionality.
         for val in val_list:
-            self.assertEqual(val, queue.pop())
+            self.assertEqual(val, q.pop())
 
     def test_stack(self):
-        stack = containers.Stack()
-        self.assertTrue(stack.isEmpty())
+        intStack = stack.Stack()
+        self.assertTrue(intStack.isEmpty())
 
         val_list = [x for x in range(1, 10)]
         for val in val_list:
-            stack.push(val)
-        self.assertFalse(stack.isEmpty())
-        self.assertEquals(len(val_list), len(stack))
+            intStack.push(val)
+        self.assertFalse(intStack.isEmpty())
+        self.assertEquals(len(val_list), len(intStack))
 
         # Test Stack for LIFO functionality.
         for val in reversed(val_list):
-            self.assertEqual(val, stack.pop())
+            self.assertEqual(val, intStack.pop())
 
     def test_priority_queue(self):
-        priority_queue = containers.PriorityQueue()
-        self.assertTrue(priority_queue.isEmpty())
+        priQ = priority_queue.PriorityQueue()
+        self.assertTrue(priQ.isEmpty())
 
         # List of values with a priority number that correspondes to the position in the list.
         val_list = [(x, -x) for x in range(1, 10)]
         for val, pri in val_list:
-            priority_queue.push(val, pri)
-        self.assertFalse(priority_queue.isEmpty())
-        self.assertEquals(len(val_list), len(priority_queue))
+            priQ.push(val, pri)
+        self.assertFalse(priQ.isEmpty())
+        self.assertEquals(len(val_list), len(priQ))
 
         # Test Min Queue for priority functionality.
         for val, pri in reversed(val_list):
-            self.assertEqual(val, priority_queue.pop())
+            self.assertEqual(val, priQ.pop())
 
 if __name__ == '__main__':
     unittest.main()
